Current Arduino Code (10/27/2021)
The current code is set to output a 10kHz 50% Duty Cycle square pulse waveform. The commented out code is preliminary code for the OLED display which doesn't behave at the moment. 
The goal is to control the frequency modulation via a rotary encoder and display it on the screen as well as the selected waveform.